Clean up support crate a bit
