+++ tests/codec_read.rs
@@ -1,6 +1,6 @@
 #[macro_use]
 extern crate h2_test_support;
-use h2_test_support::*;
+use h2_test_support::prelude::*;
 
 #[test]
 fn read_none() {
