+++ tests/prioritization.rs
@@ -1,5 +1,5 @@
 extern crate h2_test_support;
-use h2_test_support::*;
+use h2_test_support::prelude::*;
 
 #[test]
 fn single_stream_send_large_body() {
