+++ tests/stream_states.rs
@@ -2,7 +2,7 @@
 extern crate log;
 
 extern crate h2_test_support;
-use h2_test_support::*;
+use h2_test_support::prelude::*;
 
 #[test]
 fn send_recv_headers_only() {
