+++ tests/support/src/assert.rs
@@ -1,4 +1,3 @@
-// ===== Assertions =====
 
 #[macro_export]
 macro_rules! assert_closed {
@@ -48,51 +47,4 @@ macro_rules! poll_data {
             frame => panic!("expected data frame; actual={:?}", frame),
         }
     }}
-}
-
-// ===== Build a codec from raw bytes =====
-
-#[macro_export]
-macro_rules! raw_codec {
-    (
-        $(
-            $fn:ident => [$($chunk:expr,)+];
-        )*
-    ) => {{
-        let mut b = $crate::mock_io::Builder::new();
-
-        $({
-            let mut chunk = vec![];
-
-            $(
-                $crate::codec::Chunk::push(&$chunk, &mut chunk);
-            )+
-
-            b.$fn(&chunk[..]);
-        })*
-
-        $crate::Codec::new(b.build())
-    }}
-}
-
-pub trait Chunk {
-    fn push(&self, dst: &mut Vec<u8>);
-}
-
-impl Chunk for u8 {
-    fn push(&self, dst: &mut Vec<u8>) {
-        dst.push(*self);
-    }
-}
-
-impl<'a> Chunk for &'a [u8] {
-    fn push(&self, dst: &mut Vec<u8>) {
-        dst.extend(*self)
-    }
-}
-
-impl<'a> Chunk for &'a str {
-    fn push(&self, dst: &mut Vec<u8>) {
-        dst.extend(self.as_bytes())
-    }
 }
